Start to change gathering and move to more compilation from source

+++ compile_zlib.sh
@@ -4,7 +4,7 @@
 
 basedir=/src/zlib
 baseurl=http://zlib.net
-version=1.2.5
+version=1.2.7
 tarball=zlib-${version}.tar.gz
 directory=zlib-${version}
 
@@ -21,22 +21,33 @@ fi
 
 pushd $directory
 
-if [ ! -f mingw.patch.applied ]
+if [ -f ../../zlib-${version}-mingw.patch ]
 then
-    echo patching ...
-    patch -p1 < ../../zlib-${version}-mingw.patch
-    echo applied > mingw.patch.applied
+    if [ ! -f mingw.patch.applied ]
+    then
+        echo patching ...
+        patch -p1 < ../../zlib-${version}-mingw.patch
+        echo applied > mingw.patch.applied
+    fi
 fi
 
 
 make -fwin32/Makefile.gcc
-cp libz-1.dll /bin
+cp zlib1.dll /bin
 cp libz.dll.a /lib
 cp libz.a /lib
 cp zlib.h /include
 cp zconf.h /include
 
 
+# prepare gather
+cp zlib1.dll $prepare_bin
+cp ../zlib-${version}.tar.gz $prepare_src
+
+if [ -f ../../zlib-${version}-mingw.patch ]
+then
+    cp ../../zlib-${version}-mingw.patch $prepare_src
+fi
 
 
 popd
